add 'readwrite' and 'flags' options to Database.new

+++ test/test_database_readonly.rb
@@ -11,12 +11,19 @@ module SQLite3
 
     def teardown
       @db.close unless @db.closed?
-      File.unlink 'test-readonly.db'
+      File.unlink 'test-readonly.db' if File.exists?('test-readonly.db')
     end
 
     def test_open_readonly_database
       @db = SQLite3::Database.new('test-readonly.db', :readonly => true)
       assert @db.readonly?
+    end
+
+    def test_open_readonly_not_exists_database
+      File.unlink 'test-readonly.db'
+      assert_raise(SQLite3::CantOpenException) do
+        @db = SQLite3::Database.new('test-readonly.db', :readonly => true)
+      end
     end
 
     def test_insert_readonly_database
